refactored schema and respective models
