+++ PizzaParlor/Models/Customer.cs
@@ -4,14 +4,10 @@ namespace PizzaParlor.Models
 {
   public class Customer
   {
-    public Customer()
-    {
-      this.JoinEntities = new HashSet<CustomerOrder>();
-    }
-
     public int CustomerId { get; set; }
     public string Name { get; set; }
     public string ContactInfo { get; set; }
     public virtual ICollection<CustomerOrder> JoinEntities { get; set; }
+    public virtual ICollection<Order> Orders { get; set; }
   }
 }
