+++ PizzaParlor/Models/MenuItem.cs
@@ -1,13 +1,20 @@
 using System.Collections.Generic;
 
-namespace PizzaParlor
+namespace PizzaParlor.Models
 {
   public class MenuItem
   {
+    public MenuItem()
+    {
+      this.JoinEntities = new HashSet<MenuItemOrder>();
+    }
+
     public int MenuItemId { get; set; }
     public string Name { get; set; }
     public int price { get; set; }
     public string Description { get; set; }
     public bool IsTopping { get; set; }
+    public virtual ICollection<MenuItemOrder> JoinEntities { get; set; }
+
   }
 }
